+++ lib/constants/translations/app_strings.dart
@@ -1,7 +1,6 @@
 import 'package:defend_your_flame/constants/translations/app_string_constants.dart';
 import 'package:defend_your_flame/constants/translations/app_string_data.dart';
 import 'package:defend_your_flame/constants/translations/app_string_helper.dart';
-import 'package:defend_your_flame/helpers/global_vars.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';