+++ lib/core/flame/components/entities/entity.dart
@@ -4,6 +4,7 @@ import 'package:defend_your_flame/core/flame/components/entities/enums/entity_st
 import 'package:defend_your_flame/core/flame/components/entities/configs/entity_config.dart';
 import 'package:defend_your_flame/core/flame/components/entities/enums/idle_time.dart';
 import 'package:defend_your_flame/core/flame/components/entities/mixins/has_hitbox_positioning.dart';
+import 'package:defend_your_flame/core/flame/components/entities/mobs/rock_golem.dart';
 import 'package:defend_your_flame/core/flame/main_game.dart';
 import 'package:defend_your_flame/core/flame/managers/entity_manager.dart';
 import 'package:defend_your_flame/core/flame/managers/sprite_manager.dart';
@@ -29,36 +30,35 @@ class Entity extends SpriteAnimationGroupComponent<EntityState>
         WallAsSolid {
   static const double offscreenTimeoutInSeconds = 3;
 
+  late final List<ShapeHitbox> _hitboxes = addHitboxes();
+  late final double _attackOffset = entityConfig.attackRange == null ? 0 : entityConfig.attackRange!() * scale.x;
+  late final double _defaultWalkingSpeed;
+
   final EntityConfig entityConfig;
   final double scaleModifier;
 
   late double _currentHealth = entityConfig.totalHealth;
-
-  late final List<ShapeHitbox> _hitboxes = addHitboxes();
-
-  late final double _attackOffset = entityConfig.attackRange == null ? 0 : entityConfig.attackRange!() * scale.x;
-
   late Vector2 _startingPosition;
 
   bool _canAttack = false;
   double _offscreenTimerInMilliseconds = 0;
 
-  bool get isAlive => _currentHealth > MiscConstants.eps;
-
   double get currentHealth => _currentHealth;
   double get totalHealth => entityConfig.totalHealth;
+  double get walkingSpeed => _defaultWalkingSpeed;
 
   Vector2 get startPosition => _startingPosition;
-
   Vector2 get trueCenter => absoluteCenterOfMainHitbox();
 
   bool get isWalking => current == EntityState.walking;
+  bool get isAlive => _currentHealth > MiscConstants.eps;
 
   TimeSpendIdle get idleTime => entityConfig.timeSpendIdle;
 
-  Entity({required this.entityConfig, this.scaleModifier = 1}) {
+  Entity({required this.entityConfig, this.scaleModifier = 1, double? modifiedWalkingSpeed}) {
     size = entityConfig.defaultSize;
     scale = Vector2.all(entityConfig.defaultScale * scaleModifier);
+    _defaultWalkingSpeed = modifiedWalkingSpeed ?? entityConfig.baseWalkingSpeed;
   }
 
   @override
@@ -69,8 +69,11 @@ class Entity extends SpriteAnimationGroupComponent<EntityState>
 
   @override
   Future<void> onLoad() async {
+    var speedDifference = _defaultWalkingSpeed / entityConfig.baseWalkingSpeed;
     final walkingSprite = SpriteManager.getAnimation('mobs/${entityConfig.entityResourceName}/walk',
-        stepTime: entityConfig.walkingConfig.stepTime / scale.x, frames: entityConfig.walkingConfig.frames, loop: true);
+        stepTime: entityConfig.walkingConfig.stepTime / scale.x / speedDifference,
+        frames: entityConfig.walkingConfig.frames,
+        loop: true);
 
     final attackingSprite = SpriteManager.getAnimation('mobs/${entityConfig.entityResourceName}/attack',
         stepTime: entityConfig.attackingConfig.stepTime / scale.x,
@@ -85,6 +88,7 @@ class Entity extends SpriteAnimationGroupComponent<EntityState>
           stepTime: entityConfig.idleConfig!.stepTime / scale.x, frames: entityConfig.idleConfig!.frames, loop: true);
 
       animations = {
+        ...?animations,
         EntityState.idle: idleSprite,
       };
     }
@@ -180,7 +184,7 @@ class Entity extends SpriteAnimationGroupComponent<EntityState>
     if (isCollidingWithWall) {
       wallCollisionCalculation(dt);
     } else if (current == EntityState.walking) {
-      position.x = TimestepHelper.add(position.x, entityConfig.walkingForwardSpeed * scale.x, dt);
+      position.x = TimestepHelper.add(position.x, entityConfig.baseWalkingSpeed * scale.x, dt);
 
       if (_attackOffset > MiscConstants.eps) {
         final horizontalDistanceToWall = (world.playerBase.position.x - trueCenter.x).abs();
@@ -203,9 +207,10 @@ class Entity extends SpriteAnimationGroupComponent<EntityState>
   void takeDamage(double damage, {Vector2? position}) {
     _currentHealth -= damage;
 
-    world.effectManager.addDamageText(damage.toInt(), position ?? trueCenter);
     if (_currentHealth <= MiscConstants.eps) {
       initiateDeath();
+    } else {
+      world.effectManager.addDamageText(damage.toInt(), position ?? trueCenter);
     }
   }