+++ lib/core/flame/components/entities/mobs/rock_golem.dart
@@ -3,32 +3,37 @@ import 'package:defend_your_flame/core/flame/components/entities/configs/animati
 import 'package:defend_your_flame/core/flame/components/entities/disappear_on_death.dart';
 import 'package:defend_your_flame/core/flame/components/entities/draggable_entity.dart';
 import 'package:defend_your_flame/core/flame/components/entities/configs/entity_config.dart';
+import 'package:defend_your_flame/core/flame/components/entities/enums/idle_time.dart';
+import 'package:defend_your_flame/core/flame/components/entities/mixins/has_idle_time.dart';
 import 'package:defend_your_flame/core/flame/helpers/entity_helper.dart';
 import 'package:defend_your_flame/helpers/misc_helper.dart';
 import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
 import 'package:flutter/rendering.dart';
 
-class StrongSkeleton extends DraggableEntity with DisappearOnDeath {
-  static final EntityConfig _strongSkeletonConfig = EntityConfig(
-    entityResourceName: 'strong_skeleton',
-    defaultSize: Vector2(64, 64),
-    defaultScale: 1.65,
-    walkingConfig: AnimationConfig(frames: 12, stepTime: 0.15),
-    attackingConfig: AnimationConfig(frames: 13, stepTime: 0.11),
-    dragConfig: AnimationConfig(frames: 4, stepTime: 0.16),
+class RockGolem extends DraggableEntity with DisappearOnDeath, HasIdleTime {
+  static final EntityConfig _rockGolemConfig = EntityConfig(
+    entityResourceName: 'rock_golem',
+    defaultSize: Vector2(90, 64),
+    defaultScale: 1.4,
+    walkingConfig: AnimationConfig(frames: 10, stepTime: 0.15),
+    attackingConfig: AnimationConfig(frames: 11, stepTime: 0.13),
+    dragConfig: AnimationConfig(frames: 8, stepTime: 0.16),
+    idleConfig: AnimationConfig(frames: 8, stepTime: 0.15),
     dyingConfig: AnimationConfig(frames: 13, stepTime: 0.14),
     damageOnAttack: 15,
     goldOnKill: 8,
-    baseWalkingSpeed: 19,
-    totalHealth: DamageConstants.fallDamage * 2,
-    dragResistance: 0.6,
+    baseWalkingSpeed: 22,
+    totalHealth: DamageConstants.fallDamage * 4,
+    dragResistance: 0.52,
+    timeSpendIdle: TimeSpendIdle.minimal,
+    magicImmune: true,
   );
 
   late final RectangleHitbox _hitbox =
-      EntityHelper.createRectangleHitbox(size: Vector2(24, 33), position: Vector2(31, 48), anchor: Anchor.bottomCenter);
+      EntityHelper.createRectangleHitbox(size: Vector2(35, 33), position: Vector2(44, 64), anchor: Anchor.bottomCenter);
 
-  StrongSkeleton({super.scaleModifier, super.modifiedWalkingSpeed}) : super(entityConfig: _strongSkeletonConfig);
+  RockGolem({super.scaleModifier, super.modifiedWalkingSpeed}) : super(entityConfig: _rockGolemConfig);
 
   @override
   Vector2? attackEffectPosition() {
@@ -48,10 +53,10 @@ class StrongSkeleton extends DraggableEntity with DisappearOnDeath {
     super.render(canvas);
   }
 
-  static StrongSkeleton spawn({required Vector2 position, required double speedFactor}) {
+  static RockGolem spawn({required Vector2 position, required double speedFactor}) {
     final scaleModifier = MiscHelper.randomDouble(minValue: 1, maxValue: 1.2);
-    final skeleton = StrongSkeleton(scaleModifier: scaleModifier);
-    skeleton.position = position - Vector2(skeleton.scaledSize.x, skeleton.scaledSize.y / 2);
-    return skeleton;
+    final golem = RockGolem(scaleModifier: scaleModifier);
+    golem.position = position - Vector2(golem.scaledSize.x, golem.scaledSize.y / 2);
+    return golem;
   }
 }